address PR #32 review
- Drop the `?? "invalid_skill"` dead fallback in
  `src/sources/expand.ts`'s `pushSoft`. `loadSkill` only throws
  `CrewError`, whose `.code` is typed non-null; a non-CrewError
  would indicate a bug upstream and should surface loudly, not be
  silently converted into an `invalid_skill` skip.
- Drop the `wrapped[0] ?? s.message` dead fallback in the
  "Failed" section renderer. `wrap` always returns ≥1 element for
  a non-empty message.
- Drop the unreachable `"nothing to do"` fallback in
  `formatTotalsWithSkipped` — after the early-return guard, at
  least one of the counts is > 0 so `parts` is non-empty.
- Document the `info` command's behavior on skipped skills:
  `expandSkills` returns `{ valid, skipped }`; `info` deliberately
  drops skips because it's a read-only "show me what's valid here"
  command. `crew install` is the one that surfaces validation
  failures per §9 step 9.
- Split `src/commands/install/index.ts` (212 lines, over the cap)
  by extracting `promptForCollision` + `abortError` to
  `collision-prompt.ts`. Top-level command is now 155 lines.
- Add C-INST-21 test: a multi-skill dir where EVERY skill fails
  validation exits 4 (not 1) and lists both invalid skills in the
  Failed section. Covers the `valid.length === 0 && skipped.length > 0`
  path that the new `expandSkills` guard intentionally allows.
